+++ service-implementation/src/main/java/com/liaison/mailbox/service/core/processor/FTPSRemoteUploader.java
@@ -15,6 +15,7 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.util.Date;
 
+import com.liaison.mailbox.service.executor.javascript.JavaScriptExecutorUtil;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -76,6 +77,7 @@ public class FTPSRemoteUploader extends AbstractRemoteUploader {
             FTPUploaderPropertiesDTO ftpUploaderStaticProperties = (FTPUploaderPropertiesDTO)getProperties();
             
             boolean recursiveSubdirectories = ftpUploaderStaticProperties.isRecurseSubDirectories();
+            setDirectUpload(ftpUploaderStaticProperties.isDirectUpload());
             File[] subFiles = getFilesToUpload(recursiveSubdirectories);
             if (subFiles == null || subFiles.length == 0) {
                 LOGGER.info(constructMessage("The given payload location {} doesn't have files to upload."), path);
@@ -140,7 +142,7 @@ public class FTPSRemoteUploader extends AbstractRemoteUploader {
             }
 
             //FTPS
-            uploadFile(ftpsRequest, localParentDir, remoteParentDir, item);
+            uploadFile(ftpsRequest, remoteParentDir, item);
         }
     }
 
@@ -155,10 +157,11 @@ public class FTPSRemoteUploader extends AbstractRemoteUploader {
      * @throws LiaisonException
      * @throws IllegalAccessException
      */
-    private void uploadFile(G2FTPSClient ftpsRequest, String localParentDir, String remoteParentDir, File file)
+    private void uploadFile(G2FTPSClient ftpsRequest, String remoteParentDir, File file)
             throws IOException, LiaisonException, IllegalAccessException {
 
         int replyCode;
+        String filePath = file.getParent();
         FTPUploaderPropertiesDTO staticProp = (FTPUploaderPropertiesDTO) getProperties();
         String statusIndicator = staticProp.getFileTransferStatusIndicator();
         String currentFileName = file.getName();
@@ -179,7 +182,7 @@ public class FTPSRemoteUploader extends AbstractRemoteUploader {
         try (InputStream inputStream = new FileInputStream(file)) {
             ftpsRequest.changeDirectory(remoteParentDir);
             LOGGER.info(constructMessage("uploading file {} from local path {} to remote path {}"),
-                    currentFileName, localParentDir, remoteParentDir);
+                    currentFileName, filePath, remoteParentDir);
             replyCode = ftpsRequest.putFile(uploadingFileName, inputStream);
         }
 
@@ -215,7 +218,7 @@ public class FTPSRemoteUploader extends AbstractRemoteUploader {
                     .append("Failed to upload file ")
                     .append(currentFileName)
                     .append(" from local path ")
-                    .append(localParentDir)
+                    .append(filePath)
                     .append(" to remote path ")
                     .append(remoteParentDir);
 
@@ -238,6 +241,54 @@ public class FTPSRemoteUploader extends AbstractRemoteUploader {
             } catch (LiaisonException e) {
                 LOGGER.error(constructMessage("Failed to close connection"), e);
             }
+        }
+    }
+
+    @Override
+    public void doDirectUpload(String fileName, String folderPath) {
+
+        setDirectUpload(true);
+        boolean isHandOverExecutionToJavaScript = false;
+        try {
+            isHandOverExecutionToJavaScript = ((FTPUploaderPropertiesDTO) getProperties()).isHandOverExecutionToJavaScript();
+        } catch (IllegalArgumentException | IllegalAccessException | IOException e) {
+            throw new RuntimeException(e);
+        }
+
+        if (isHandOverExecutionToJavaScript) {
+            setFileName(fileName);
+            setFolderPath(folderPath);
+            JavaScriptExecutorUtil.executeJavaScript(configurationInstance.getJavaScriptUri(), this);
+        } else {
+
+            G2FTPSClient ftpsRequest = null;
+            try {
+
+                String remotePath = validateRemotePath();
+
+                ftpsRequest = (G2FTPSClient) getClient();
+                ftpsRequest.setLogPrefix(constructMessage());
+                ftpsRequest.connect();
+                ftpsRequest.login();
+
+                // retrieve required properties
+                FTPUploaderPropertiesDTO ftpUploaderStaticProperties = (FTPUploaderPropertiesDTO) getProperties();
+                setPassive(ftpsRequest, ftpUploaderStaticProperties);
+
+                changeDirectory(ftpsRequest, remotePath);
+                LOGGER.info(constructMessage("Ready to upload file {} from local path {} to remote path {}"),
+                        fileName,
+                        folderPath,
+                        remotePath);
+                uploadFile(ftpsRequest, remotePath, new File(folderPath + File.separatorChar + fileName));
+
+            } catch (Exception e) {
+                LOGGER.error(constructMessage("Error occurred during direct upload", seperator, e.getMessage()), e);
+                throw new RuntimeException(e);
+            } finally {
+                disconnect(ftpsRequest);
+            }
+
         }
     }